refactor: Node Component - for proper hook usage
Refactored the BreadCrumb component to adhere to the 'Rules of Hooks'. Although useFetch() was being utilized as a regular function, its naming convention suggested it was a hook, which could lead to confusion and potential misuse. To avoid violating the rules of Hooks, which state that hooks shouldn't be called inside event handlers, useFetch() was removed from the BreadCrumb component. Instead, the component now employs an onPathChange() function to handle path updates.